+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## Unreleased
+
+## 0.5.0 (6. September, 2023)
 ### Changed
 - (breaking) SessionStore::initiate() is removed. initiate has been merged into SessionStore::new(). 
 ### Fixed